+++ lib/src/widgets/animation/animated_ghost.dart
