+++ lib/src/widgets/misc/size_overlay.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:v_flutter_core/src/widgets/size_reporter.dart';
+import 'package:v_flutter_core/src/widgets/misc/size_reporter.dart';
 
 class SizeOverlay extends HookWidget {
   const SizeOverlay({
