+++ lib/src/widgets/progress_button/progress_button.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:v_flutter_core/src/widgets/progress_button/button_variants.dart';
 import 'package:v_flutter_core/src/widgets/progress_button/delegating_button_style_button.dart';
-import 'package:v_flutter_core/src/widgets/size_reporter.dart';
+import 'package:v_flutter_core/src/widgets/misc/size_reporter.dart';
 
 class ProgressButton extends HookWidget {
   const ProgressButton.variant({
