+++ lib/src/widgets/progress_button/progress_button_indicator.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:v_flutter_core/src/widgets/size_reporter.dart';
+import 'package:v_flutter_core/src/widgets/misc/size_reporter.dart';
 
 class ProgressButtonIndicator extends StatelessWidget {
   const ProgressButtonIndicator({
