+++ lib/v_flutter_core.dart
@@ -41,35 +41,38 @@ export 'src/utils/type_checks.dart';
 
 export 'src/widgets/animation/animated_collapse.dart';
 export 'src/widgets/animation/animated_collapse_and_opacity.dart';
+export 'src/widgets/animation/animated_ghost.dart';
 export 'src/widgets/animation/size_then_fade_transition.dart';
 export 'src/widgets/dropdown/cascading_dropdown.dart';
 export 'src/widgets/dropdown/cascading_dropdown_theme.dart';
 export 'src/widgets/dropdown/step_focus.dart';
-export 'src/widgets/flex/gap_column.dart';
-export 'src/widgets/flex/gap_flex.dart';
-export 'src/widgets/flex/gap_row.dart';
-export 'src/widgets/flex/separated_column.dart';
-export 'src/widgets/flex/separated_flex.dart';
-export 'src/widgets/flex/separated_row.dart';
 export 'src/widgets/inheriting_scaffold/delegating_scaffold.dart';
 export 'src/widgets/inheriting_scaffold/inheriting_scaffold.dart';
 export 'src/widgets/inheriting_scaffold/scaffold_behavior.dart';
 export 'src/widgets/layout/adaptive_width_wrap.dart';
+export 'src/widgets/layout/constraint_adaptive_layout.dart';
+export 'src/widgets/layout/flex/gap_column.dart';
+export 'src/widgets/layout/flex/gap_flex.dart';
+export 'src/widgets/layout/flex/gap_row.dart';
+export 'src/widgets/layout/flex/separated_column.dart';
+export 'src/widgets/layout/flex/separated_flex.dart';
+export 'src/widgets/layout/flex/separated_row.dart';
+export 'src/widgets/layout/size_class_layout.dart';
 export 'src/widgets/misc/always_scrollbar.dart';
-export 'src/widgets/misc/animated_ghost.dart';
-export 'src/widgets/misc/constraint_adaptive_layout.dart';
 export 'src/widgets/misc/debug_container.dart';
 export 'src/widgets/misc/decorated_single_child_scroll_view.dart';
 export 'src/widgets/misc/dismissible_portal_target.dart';
 export 'src/widgets/misc/ghost.dart';
 export 'src/widgets/misc/maybe_align.dart';
 export 'src/widgets/misc/portal_dialog.dart';
-export 'src/widgets/misc/size_class_layout.dart';
+export 'src/widgets/misc/seamless_circular_progress_indicator.dart';
 export 'src/widgets/misc/size_overlay.dart';
+export 'src/widgets/misc/size_reporter.dart';
 export 'src/widgets/misc/tab_visibility_change_recognizer.dart';
 export 'src/widgets/misc/target_platform_widget.dart';
 export 'src/widgets/misc/uninteractable.dart';
 export 'src/widgets/misc/wide.dart';
+
 export 'src/widgets/progress_button/button_variants.dart';
 export 'src/widgets/progress_button/progress_button.dart';
 export 'src/widgets/reactive_text_field/autocomplete/autocomplete_decoration.dart';
@@ -84,8 +87,6 @@ export 'src/widgets/reactive_text_field/reactive_text_field_theme.dart';
 export 'src/widgets/reactive_text_field/themed_reactive_text_field.dart';
 export 'src/widgets/reactive_text_field/typed_delegate_validator.dart';
 export 'src/widgets/reactive_text_field/utils.dart';
-export 'src/widgets/seamless_circular_progress_indicator.dart';
-export 'src/widgets/size_reporter.dart';
 export 'src/widgets/type_scale/type_scale_text.dart';
 export 'src/widgets/type_scale/type_scale_variant.dart';
 export 'src/widgets/type_scale/underline_default_text_style.dart';
